Move responsibility for managing ES into `h.api.search`
This commit moves management of the ElasticSearch index settings,
mappings, etc, into `h.api.search`.

This further decouples the search package from annotator-store, as we
now manage the index mappings and analysis settings ourselves (which we
were very nearly doing, anyway) rather than relying on some parts of
annotator-store.

I've also removed `h.reindexer` in favour of about 3 lines of code that
uses `elasticsearch.helpers.reindex` to achieve much the same end, with
the added bonus that we now use exactly the same code to configure the
new indices as we use to configure indices on startup.
