+++ conf/development.ini
@@ -6,6 +6,7 @@ pipeline: h
 use: egg:h
 
 h.db.should_create_all: True
+h.search.autoconfigure: True
 
 es.host: http://localhost:9200
