+++ h/api/models/test/elastic_test.py
@@ -1,68 +1,10 @@
 # -*- coding: utf-8 -*-
-# pylint: disable=no-self-use
-import itertools
-import re
-import urllib
 
 import pytest
 from mock import patch
 from mock import PropertyMock
 
 from h.api.models.elastic import Annotation
-
-analysis = Annotation.__analysis__
-
-
-def test_strip_scheme_char_filter():
-    f = analysis['char_filter']['strip_scheme']
-    p = f['pattern']
-    r = f['replacement']
-    assert(re.sub(p, r, 'http://ping/pong#hash') == 'ping/pong#hash')
-    assert(re.sub(p, r, 'chrome-extension://1234/a.js') == '1234/a.js')
-    assert(re.sub(p, r, 'a+b.c://1234/a.js') == '1234/a.js')
-    assert(re.sub(p, r, 'uri:x-pdf:1234') == 'x-pdf:1234')
-    assert(re.sub(p, r, 'example.com') == 'example.com')
-    # This is ambiguous, and possibly cannot be expected to work.
-    # assert(re.sub(p, r, 'localhost:5000') == 'localhost:5000')
-
-
-def test_path_url_filter():
-    patterns = analysis['filter']['path_url']['patterns']
-    assert(captures(patterns, 'example.com/foo/bar?query#hash') == [
-        'example.com/foo/bar'
-    ])
-    assert(captures(patterns, 'example.com/foo/bar/') == [
-        'example.com/foo/bar/'
-    ])
-
-
-def test_rstrip_slash_filter():
-    p = analysis['filter']['rstrip_slash']['pattern']
-    r = analysis['filter']['rstrip_slash']['replacement']
-    assert(re.sub(p, r, 'example.com/') == 'example.com')
-    assert(re.sub(p, r, 'example.com/foo/bar/') == 'example.com/foo/bar')
-
-
-def test_uri_part_tokenizer():
-    text = 'http://a.b/foo/bar?c=d#stuff'
-    pattern = analysis['tokenizer']['uri_part']['pattern']
-    assert(re.split(pattern, text) == [
-        'http', '', '', 'a', 'b', 'foo', 'bar', 'c', 'd', 'stuff'
-    ])
-
-    text = urllib.quote_plus(text)
-    assert(re.split(pattern, 'http://jump.to/?u=' + text) == [
-        'http', '', '', 'jump', 'to', '', 'u',
-        'http', '', '', 'a', 'b', 'foo', 'bar', 'c', 'd', 'stuff'
-    ])
-
-
-def captures(patterns, text):
-    return list(itertools.chain(*(groups(p, text) for p in patterns)))
-
-
-def groups(pattern, text):
-    return re.search(pattern, text).groups() or []
 
 
 def test_uri():
